Fix crash in lastupdate() method
When a RRD database contains floating point values, the
PyRRD_Int_FromString function returns NULL, which will then be tried
to be put into a dictionary.
This operation fails as PyDict_SetItemString does not handle NULL
and the program crashes with SEGV.

This patch parses the value as a floating point number
and will add that instead. If the value could not be parsed,
None will be added instead.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>